requirements.txt
scipy==1.4.1
h5py
cython
pytz
opencv-python
tqdm
pyyaml
tensorboard
tabulate
termcolor
joblib
click
atari-py
tensorflow_gpu==2.1.0
cycler
kiwisolver
python-dateutil
pyparsing
cloudpickle==1.2.2
pyglet<=1.3.2
tensorboardX
scikit-learn
yacs
matplotlib
easydict
pillow==6.2.1
scikit-image
pandas
packaging
torch==1.3.0
torchvision==0.4.1
